BUGFIX: Shutdown now works fine
The `.shutdown()` method used to have an `.await_idle()` call in there,
leftover from when it was necessary to wait for Multiwfn to be
ready to accept input, and then supply something invalid.

Now that shutdown is accomplished via direct kill of the process,
there's no need to wait for idle status!